services/requirements/base.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
defusedxml==0.7.1
Django==5.0.6
django-cors-headers==4.3.1
django-createsuperuser==2020.12.3
django-filter==24.2
djangorestframework==3.15.1
httpie==3.2.2
idna==3.7
Markdown==3.6
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.5
psycopg==3.1.12
psycopg-binary==3.1.12
Pygments==2.18.0
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.2
requests-toolbelt==1.0.0
rich==13.7.1
sqlparse==0.5.0
typing_extensions==4.12.0
urllib3==2.2.1

# Django

# djangorestframework
# django-createsuperuser
# markdown       # Markdown support for the browsable API.
# django-filter  # Filtering support
# python-dotenv

# asgiref
# sqlparse
# psycopg[binary]==3.1.12

# httpie

# django-cors-headers #

# # ## for predictions
# # transformers
# # # https://download.pytorch.org/whl/cpu/torch-2.0.1%2Bcpu-cp311-cp311-linux_x86_64.whl
# # # https://download.pytorch.org/whl/cpu/torch-2.1.0%2Bcpu-cp38-cp38-linux_x86_64.whl
# # https://download.pytorch.org/whl/cpu/torch-2.0.1%2Bcpu-cp310-cp310-linux_x86_64.whl